+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ---
 
+## [Unreleased] - 2026-03-11
+### Added:
+* Course UI - Added CKForm
+
+---
+
 ## [1.21.4] - 2026-03-11
 ### Fixed:
 * Deploy config